requirements.txt
annotated-types==0.7.0
anyio==4.6.0
awsebcli==3.21.0
beautifulsoup4==4.12.3
blessed==1.20.0
blinker==1.9.0
botocore==1.35.27
bs4==0.0.2
cement==2.10.14
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-MySQLdb==2.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
MarkupSafe==2.1.5
mysqlclient==2.2.6
numpy==2.1.1
openai==1.48.0
pathspec==0.10.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
scipy==1.14.1
semantic-version==2.10.0
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
termcolor==2.4.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==1.26.20
wcwidth==0.2.13
Werkzeug==3.1.3
Flask-SocketIO==5.5.1
dnspython==2.7.0
eventlet==0.39.0
greenlet==3.1.1
cffi==1.17.1
cryptography==44.0.1
jwt==1.3.1
pycparser==2.22
dotenv==0.9.9
python-dotenv==1.0.1